+++ CHANGES.md
@@ -35,6 +35,8 @@ Release Notes.
 - Property: Introduce the schema style to the property.
 - Add time range parameters to stream index filter.
 - UI: Add the `stages` to groups.
+- Add time range return value from stream local index filter.
+- Deduplicate the documents on building the series index.
 
 ### Bug Fixes
 
@@ -56,6 +58,7 @@ Release Notes.
 - UI: Update BanyanDB UI to Integrate New Property Query API.
 - UI: Fix the Stream List.
 - Fix the oom issue when loading too many unnecessary parts into memory.
+- bydbctl: Fix the bug that the bydbctl can't parse the absolute time flag.
 
 ### Documentation